client/requirements.txt
dnspython==2.4.2
colorama==0.4.6
click==8.1.7